Add files via upload
In the video 'Business Value Dashboard - Tips & Tricks' the files basicGenerator_integer_sequential.js and config_integer_sequential.json are used. The files are modifications of the files basicGenerator.js and config_.json. Please use the files in order to walk through the tasks shown in the video on your own. Please rename config_integer_sequential.json to config.json and follow the instructions mentioned below.